fix page.js and move useCE
