+++ src/app/events/form/page.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function page() {
+  return (
+    <div>page</div>
+  )
+}
