+++ src/app/page.js
@@ -1,5 +1,5 @@
 import { CreateEvent } from "./components/auth/components/CreateEvent";
 
 export default function Home() {
-  return <Login />;
+  return <div>Login</div>;
 }
